test(native-trading): co-locate screens and utilities
