+++ suite-native/module-trading/src/screens/TradingExchangeApprovalScreen.test.tsx
@@ -7,18 +7,16 @@ import {
     RootStackRoutes,
     useNavigationRemoveActionInterceptor,
 } from '@suite-native/navigation';
-import { type TestStore } from '@suite-native/test-utils-store';
+import { type TestStore, fireEvent } from '@suite-native/test-utils-store';
 import { eth1NormalAccount, mercuryoFixedWorstQuote } from '@suite-native/trading-fixtures';
 
-import {
-    createTradingLightStore,
-    renderWithTradingProvider,
-} from '../../__tests__/tradingTestUtils';
-import { TradingExchangeRevokeScreen } from '../TradingExchangeRevokeScreen';
+import { TradingExchangeApprovalScreen } from './TradingExchangeApprovalScreen';
+import { createTradingLightStore, renderWithTradingProvider } from '../__tests__/tradingTestUtils';
 
 const mockShowSheet = jest.fn();
 const mockHideSheet = jest.fn();
 const mockConfirmApproval = jest.fn().mockResolvedValue({});
+const mockOnApprovalTypeChange = jest.fn();
 const mockNavigationDispatch = jest.fn();
 
 jest.mock('@suite-native/navigation', () => ({
@@ -36,17 +34,18 @@ const triggerPreventNavigationRemove = (action: NavigationAction = { type: 'GO_B
     params?.onInterceptedAction?.(action);
 };
 
-jest.mock('../../hooks/exchange/Approval/useApprovalFlow', () => ({
+jest.mock('../hooks/exchange/Approval/useApprovalFlow', () => ({
     useApprovalFlow: () => ({
         quote: undefined,
         isReady: true,
         isConfirming: false,
         error: null,
         confirmApproval: mockConfirmApproval,
+        onApprovalTypeChange: mockOnApprovalTypeChange,
     }),
 }));
 
-jest.mock('../../hooks/exchange/Approval/useEvmApprovalFees', () => ({
+jest.mock('../hooks/exchange/Approval/useEvmApprovalFees', () => ({
     useEvmApprovalFees: () => ({
         fee: '100000',
         isLoading: false,
@@ -59,13 +58,22 @@ jest.mock('@react-navigation/native', () => ({
     ...jest.requireActual('@react-navigation/native'),
     useRoute: () =>
         ({
-            key: RootStackRoutes.TradingExchangeRevoke,
-            name: RootStackRoutes.TradingExchangeRevoke,
+            key: RootStackRoutes.TradingExchangeApproval,
+            name: RootStackRoutes.TradingExchangeApproval,
             params: {},
-        }) as RouteProp<RootStackParamList, RootStackRoutes.TradingExchangeRevoke>,
+        }) as RouteProp<RootStackParamList, RootStackRoutes.TradingExchangeApproval>,
     useNavigation: () => ({
         setOptions: jest.fn(),
     }),
+}));
+
+const mockAnalyticsReport = jest.fn();
+jest.mock('@suite-native/trading-analytics', () => ({
+    ...jest.requireActual('@suite-native/trading-analytics'),
+    useExchangeAnalyticsStepReport:
+        (action: unknown) =>
+        (...args: unknown[]) =>
+            mockAnalyticsReport(action, ...args),
 }));
 
 jest.mock('@suite-native/trading-atoms', () => ({
@@ -83,18 +91,9 @@ jest.mock('@suite-common/device', () => ({
     selectIsDeviceConnected: () => mockIsDeviceConnected,
 }));
 
-const mockAnalyticsReport = jest.fn();
-jest.mock('@suite-native/trading-analytics', () => ({
-    ...jest.requireActual('@suite-native/trading-analytics'),
-    useExchangeAnalyticsStepReport:
-        (action: unknown) =>
-        (...args: unknown[]) =>
-            mockAnalyticsReport(action, ...args),
-}));
-
 const testQuote = mercuryoFixedWorstQuote;
 
-describe('TradingExchangeRevokeScreen', () => {
+describe('TradingExchangeApprovalScreen', () => {
     let store: TestStore;
     let unmount: (() => void) | undefined;
 
@@ -102,7 +101,7 @@ describe('TradingExchangeRevokeScreen', () => {
 
     const renderScreen = (params: Record<string, unknown> = {}) => {
         const result = renderWithTradingProvider(
-            <TradingExchangeRevokeScreen
+            <TradingExchangeApprovalScreen
                 route={{ params } as any}
                 navigation={{ dispatch: mockNavigationDispatch } as any}
             />,
@@ -131,16 +130,16 @@ describe('TradingExchangeRevokeScreen', () => {
         }
     });
 
-    it('should confirm revoke with ZERO approval type', () => {
+    it('should confirm approval on mount', () => {
         renderScreen();
 
         expect(mockConfirmApproval).toHaveBeenCalledTimes(1);
         expect(mockConfirmApproval).toHaveBeenCalledWith(
-            expect.objectContaining({ approvalType: 'ZERO' }),
+            expect.objectContaining({ approvalType: 'MINIMAL' }),
         );
     });
 
-    it('should render the revoke screen with quote details', () => {
+    it('should render the approval screen with quote details', () => {
         const { getByText } = renderScreen();
 
         expect(getByText('ETH Account #1')).toBeOnTheScreen();
@@ -152,6 +151,17 @@ describe('TradingExchangeRevokeScreen', () => {
         const { getByText } = renderScreen();
 
         expect(getByText('Mercuryo')).toBeOnTheScreen();
+    });
+
+    it('should open bottom sheet when limit row is pressed', async () => {
+        const { findByText } = renderScreen();
+
+        const pressableElement = await findByText(
+            getTranslation('moduleTrading.tradingExchangeApprovalScreen.limitLabel'),
+        );
+
+        fireEvent.press(pressableElement);
+        expect(mockShowSheet).toHaveBeenCalledTimes(1);
     });
 
     it('should render continue button', () => {
@@ -166,11 +176,13 @@ describe('TradingExchangeRevokeScreen', () => {
         const { getByText, queryByText } = renderScreen();
 
         expect(
-            getByText(getTranslation('moduleTrading.tradingExchangeRevokeScreen.revokeErrorAlert')),
+            getByText(
+                getTranslation('moduleTrading.tradingExchangeApprovalScreen.approveErrorAlert'),
+            ),
         ).toBeOnTheScreen();
         expect(queryByText(getTranslation('generic.buttons.continue'))).toBeNull();
         expect(errorSpy).toHaveBeenCalledTimes(1);
-        expect(errorSpy).toHaveBeenCalledWith('No quote to revoke approval');
+        expect(errorSpy).toHaveBeenCalledWith('No quote to confirm approval');
     });
 
     it('should clear selected quote on back navigation', () => {
@@ -186,12 +198,12 @@ describe('TradingExchangeRevokeScreen', () => {
         expect(mockNavigationDispatch).toHaveBeenCalledWith(backAction);
     });
 
-    it('should render low limit info alert when shouldIncreaseLimit is true', () => {
-        const { getByText } = renderScreen({ shouldIncreaseLimit: true });
+    it('should render revoke success alert when isRevoked is true', () => {
+        const { getByText } = renderScreen({ isRevoked: true });
 
         expect(
             getByText(
-                getTranslation('moduleTrading.tradingExchangeRevokeScreen.lowLimitInfoAlert'),
+                getTranslation('moduleTrading.tradingExchangeApprovalScreen.revokeSuccessAlert'),
             ),
         ).toBeOnTheScreen();
     });
@@ -207,20 +219,20 @@ describe('TradingExchangeRevokeScreen', () => {
     });
 
     describe('analytics', () => {
-        it('should report revoke-preview visit ', () => {
+        it('should report approval-preview visit ', () => {
             renderScreen();
 
-            expect(mockAnalyticsReport).toHaveBeenCalledWith('revoke-preview', 'visit');
+            expect(mockAnalyticsReport).toHaveBeenCalledWith('approval-preview', 'visit');
             expect(mockAnalyticsReport).toHaveBeenCalledTimes(1);
         });
 
-        it('should report revoke-preview cancel on back navigation', () => {
+        it('should report approval-preview cancel on back navigation', () => {
             store.dispatch(tradingExchangeActions.saveSelectedQuote(testQuote));
             renderScreen();
 
             triggerPreventNavigationRemove({ type: 'GO_BACK' });
 
-            expect(mockAnalyticsReport).toHaveBeenCalledWith('revoke-preview', 'cancel');
+            expect(mockAnalyticsReport).toHaveBeenCalledWith('approval-preview', 'cancel');
         });
     });
 });
